randomizing bear profile picture
