+++ client/src/index.tsx
@@ -42,6 +42,23 @@ function isAuthenticated() {
   }
 }
 
+const profilePictures = [
+  "/profile_bear/profile_bear_dark_blue.svg",
+  "/profile_bear/profile_bear_light_blue.svg",
+  "/profile_bear/profile_bear_light_pink.svg",
+  "/profile_bear/profile_bear_mint.png",
+  "/profile_bear/profile_bear_orange.svg",
+  "/profile_bear/profile_bear_purple.svg",
+  "/profile_bear/profile_bear_red.svg",
+  "/profile_bear/profile_bear_yellow.svg",
+];
+
+function randomPicture() {
+  return profilePictures[Math.floor(Math.random() * profilePictures.length)];
+}
+
+const profilePicture = randomPicture();
+
 const defaultProtectedRouteProps: ProtectedRouteProps = {
   isAuthenticated: isAuthenticated(),
   authenticationPath: "/",
@@ -51,7 +68,7 @@ render(
   <BrowserRouter>
     <div className="container-fluid full-height">
       <Switch>
-        <Route name="app" exact path="/" component={App} />
+        <Route name="app" exact path="/" component={() => <App imgSrc={profilePicture}/>} />
         <Route name="admin" exact path="/admin" component={Login} />
         <Route
           name="permalink"
@@ -70,7 +87,7 @@ render(
           {...defaultProtectedRouteProps}
           exact={true}
           path="/profile"
-          component={Profile}
+          component={() => <Profile imgSrc={profilePicture} />}
         />
         <Route component={NotFound} />
       </Switch>
